Tests/LibWeb/Text/expected/css/style-invalidation/has-non-subject-universal-argument-invalidation.txt

FAIL: appends without list items recomputed 5020 elements
with list item: rgb(9, 9, 9)
list item removed: rgb(1, 1, 1)
